feat: create auth routes for user authentication

+++ src/users/users.service.ts
@@ -1,9 +1,12 @@
 import {
   ConflictException,
+  ForbiddenException,
   Injectable,
   NotFoundException,
 } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { TokenPayloadDto } from 'src/auth/dto/token-payload.dto';
+import { HashingService } from 'src/auth/hashing/hashing.service';
 import { Repository } from 'typeorm';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
@@ -14,7 +17,8 @@ export class UsersService {
   constructor(
     @InjectRepository(User)
     private readonly usersRepository: Repository<User>,
-  ) {}
+    private readonly hashingService: HashingService
+  ) { }
 
   throwNotFoundError() {
     throw new NotFoundException('USER_NOT_FOUND');
@@ -32,11 +36,12 @@ export class UsersService {
   }
 
   async create(createUserDto: CreateUserDto) {
-    // const hashedPassword = await bcrypt.hash(createUserDto.password, 10);
+    const hashedPassword = await this.hashingService.hash(createUserDto.password);
 
     try {
       const userData = {
         ...createUserDto,
+        password: hashedPassword,
         createdAt: new Date(),
         updatedAt: new Date(),
       };
@@ -53,22 +58,34 @@ export class UsersService {
     }
   }
 
-  async update(id: string, updateUserDto: UpdateUserDto) {
-    const user = await this.usersRepository.preload({
-      id,
+  async update(id: string, updateUserDto: UpdateUserDto, tokenPayload: TokenPayloadDto) {
+    const useData = {
       ...updateUserDto,
       isConfirmed: updateUserDto?.isConfirmed ?? false,
       updatedAt: new Date(),
+    }
+
+    if (updateUserDto.password) {
+      const hashedPassword = await this.hashingService.hash(updateUserDto.password)
+      useData['password'] = hashedPassword
+    }
+
+    const user = await this.usersRepository.preload({
+      id,
+      ...useData
     });
 
     if (!user) return this.throwNotFoundError();
+    if (user.id !== tokenPayload?.sub) throw new ForbiddenException('DONT_HAVE_PERMISSION')
+
     await this.usersRepository.save(user);
     return user;
   }
 
-  async remove(id: string) {
-    const user = await this.findOne(id);
+  async remove(id: string, tokenPayload: TokenPayloadDto) {
+    const user = await this.usersRepository.findOneBy({ id });
     if (!user) return this.throwNotFoundError();
+    if (id !== tokenPayload?.sub) throw new ForbiddenException('DONT_HAVE_PERMISSION')
     await this.usersRepository.remove(user);
   }
 }
